skintone detection/skin_tone_module/requirements.txt
torch>=1.12
torchvision>=0.13
opencv-python
Pillow
scikit-learn
pandas
matplotlib
jupyter
tqdm
skin-tone-classifier[all]
fastapi
uvicorn[standard]
python-multipart
tqdm
scikit-image
